Add files via upload
FIxedLengthSubnetMaskCalculator is a class which permits you to divide a nettwork into x sub networks (x > 1). To do that, you simply have to specify the network address, the subnet mask and th number of sub networks that you desire.
GetNetHost is a class which permits you to determine the network address, the broadcast address, the number of machines, the range of ip addresses from an ip address and its subnet.
menu is a function which is responsible for displaying the menu of this program.
main is the file which is going to be executed by you. It will call all the methods and functions necessary to the execution of the program.